Inline resolve_shipped_path into uv_tool.py; trim redundant comments in plugin.py
resolve_shipped_path had a single caller left (with_shipped_constraints) after
builtin_help_topics reverted to its own inline resolution, so the shared-util
rationale was gone. Inline it as a private _resolve_shipped_path in uv_tool.py
(anchored on this file's own depth), delete utils/paths.py + paths_test.py, and
move the three resolution tests into uv_tool_test.py.

Also drop plugin.py comments that just restated the next few lines (section-header
"Report results" / "Resolve package names" / "Verify all packages" comments, and
the reiterating halves of the two constraints comments), keeping the genuine "why"
comments (arg-validation ordering, git-diffing rationale). The abort-on-missing
detail now lives only in the with_shipped_constraints docstring.
